+++ src/PimEnterprise/Bundle/WorkflowBundle/Doctrine/ORM/Repository/ProductDraftRepository.php
@@ -58,7 +58,7 @@ class ProductDraftRepository extends EntityRepository implements ProductDraftRep
         $qb
             ->join('p.product', 'product')
             ->leftJoin('product.categories', 'category')
-            ->innerJoin('PimEnterpriseSecurityBundle:CategoryAccess', 'a', 'WITH', 'a.category = category')
+            ->innerJoin('PimEnterpriseSecurityBundle:ProductCategoryAccess', 'a', 'WITH', 'a.category = category')
             ->where(
                 $qb->expr()->eq('a.ownProducts', true)
             )
@@ -104,7 +104,7 @@ class ProductDraftRepository extends EntityRepository implements ProductDraftRep
 
             $qb
                 ->leftJoin('product.categories', 'category')
-                ->innerJoin('PimEnterpriseSecurityBundle:CategoryAccess', 'a', 'WITH', 'a.category = category')
+                ->innerJoin('PimEnterpriseSecurityBundle:ProductCategoryAccess', 'a', 'WITH', 'a.category = category')
                 ->where(
                     $qb->expr()->eq('a.ownProducts', true)
                 )